+++ src/Parsley/OverloadedQuotesPlugin/Plugin.hs
@@ -12,7 +12,7 @@ module Parsley.OverloadedQuotesPlugin.Plugin (plugin) where
 import Data.Generics (GenericT, GenericQ, mkT, mkQ, everywhere, gmapT)
 import GHC.Generics  (Generic)
 
-import Parsley.PluginUtils (lookupModuleInPackage, lookupName, lookupNames)
+import Parsley.PluginUtils (lookupModule, lookupNames)
 
 #if __GLASGOW_HASKELL__ >= 900
 import GHC.Driver.Plugins       (Plugin (..), defaultPlugin, purePlugin)
@@ -58,27 +58,29 @@ plugin = defaultPlugin { renamedResultAction = overloadedQuotes, pluginRecompile
 data QOps a = QOps {
     _code :: a,
     _val :: a,
-    makeQ :: a
+    makeQ :: a,
+    mkCode :: a,
+    mkVal :: a
   } deriving (Functor, Foldable, Traversable, Generic)
 
 quapplicativeStrings :: QOps String
 quapplicativeStrings = QOps {
     _code = "_code",
     _val  = "_val",
-    makeQ = "makeQ"
+    makeQ = "makeQ",
+    mkCode = "mkCode",
+    mkVal = "mkVal"
   }
 
 overloadedQuotes :: [GHC.CommandLineOption] -> TcGblEnv -> GHC.HsGroup GHC.GhcRn -> TcM (TcGblEnv, GHC.HsGroup GHC.GhcRn)
 overloadedQuotes _ gEnv rn = do
   hscEnv <- GHC.getTopEnv
-  parsley <- lookupModuleInPackage hscEnv "parsley" "Parsley.Internal.Common.Utils"
+  parsley <- lookupModule hscEnv "Parsley.Internal.Bridge"
   qops <- lookupNames parsley quapplicativeStrings
-  prelude <- lookupModuleInPackage hscEnv "base" "GHC.Err"
-  undef <- lookupName prelude "undefined"
   -- This is a little inefficient, since the top-down transformation means no quotes can be
   -- found under a top-level one: we use a top-down version of everywhereBut to stop traversal
   -- of this whenever it fires
-  return (gEnv, onlyTopmost (mkQ False isQuote) (mkT (transformUTHQuote qops undef)) rn)
+  return (gEnv, onlyTopmost (mkQ False isQuote) (mkT (transformUTHQuote qops)) rn)
 
 mkApp :: GHC.SrcSpan -> Expr -> Expr -> Expr
 mkApp s f = GHC.L s . Expr.HsApp noExt f . mkPar s
@@ -103,17 +105,19 @@ isQuote _                   = False
 -- the original quote. Perhaps in that case we could just inline the definition into both holes...
 -- As `transform` works bottom up, we can always assume nested quotes are already handled: this might
 -- get tricky, however.
-transformUTHQuote :: QOps GHC.Name -> GHC.Name -> Expr -> Expr
-transformUTHQuote ops undef (LUTHQuote s ex ex' x) = --pprTouch "new quote" $
+transformUTHQuote :: QOps GHC.Name -> Expr -> Expr
+transformUTHQuote ops (LUTHQuote s ex ex' x) = --pprTouch "new quote" $
   mkPar s (makeQS `mkAppS` everywhere (mkT (transformUTHQuoteVar (_val ops) makeVal)) x
                   `mkAppS` mkQuote (everywhere (mkT (transformUTHQuoteCode (_code ops) makeCode)) x))
   where
     mkQuote y = GHC.L s (Expr.HsBracket ex (Expr.TExpBr ex' y))
     mkAppS = mkApp s
     makeQS = mkVar s (makeQ ops)
-    makeVal y = mkPar s (makeQS `mkAppS` y `mkAppS` mkVar s undef)
-    makeCode y = mkPar s (makeQS `mkAppS` mkVar s undef `mkAppS` y)
-transformUTHQuote _ _ x = x
+    mkValS = mkVar s (mkVal ops)
+    mkCodeS = mkVar s (mkCode ops)
+    makeVal y = mkPar s (mkValS `mkAppS` y)
+    makeCode y = mkPar s (mkCodeS `mkAppS` y)
+transformUTHQuote _ x = x
 
 transformUTHQuoteVar :: GHC.Name -> (Expr -> Expr) -> Expr -> Expr
 transformUTHQuoteVar _    makeVal (LUTHQuote _ _ _ x)      = makeVal x