+++ test/src/Main.hs
@@ -3,16 +3,17 @@ module Main(main) where
 
 import A
 import Data.Functor.Identity
-import Parsley (_code, WQ)
+import Parsley.Internal (_code, WQ)
 
 main = do
-  print ($$(_code @WQ test3) 'a')
-  print ((runIdentity test4) 'a')
-  print (just 'a')
-  print (qux 'a')
+  putStrLn "Testing"
+  --print ($$(_code @WQ test3) 'a')
+  --print ((runIdentity test4) 'a')
+  --print (just 'a')
+  --print (qux 'a')
 
-just :: a -> Maybe a
-just = $$(_code @WQ test7)
+--just :: a -> Maybe a
+--just = $$(_code @WQ test7)
 
-qux :: a -> a
-qux = $$(_code @WQ test5)
+--qux :: a -> a
+--qux = $$(_code @WQ test5)