+++ ChangeLog.md
@@ -7,3 +7,8 @@
 ## 0.1.0.1 -- 2021-06-11
 
 * Support for GHC 9.0 in place.
+
+## 0.1.1.0 -- 2021-06-12
+
+* Switched to support `parsley-core` as the base
+* Introduced the `Parsley.Internal.Bridge` module to provide functions for the plugin to latch onto